+++ mono_state/lib/src/action.dart
@@ -1,7 +1,19 @@
 class Action {
   final String type;
-  Action({required this.type});
+  Action({this.type = ''});
 
   @override
   String toString() => 'Action(type: $type)';
+}
+
+class RegisterStateAction extends Action {
+  RegisterStateAction(String type) : super(type: type);
+}
+
+class UnregisterStateAction extends Action {
+  UnregisterStateAction(String type) : super(type: type);
+}
+
+class ImportStateAction extends Action {
+  ImportStateAction(String type) : super(type: type);
 }