+++ mono_state/lib/src/monoState.dart
@@ -6,15 +6,15 @@ import 'stateBase.dart';
 
 class MonoState {
   BehaviorSubject<Action>? __dispatcher;
-  BehaviorSubject<Map<String, dynamic>>? __store;
-  Map<String, StreamSubscription<Action>>? __subscriptions;
+  BehaviorSubject<Map<Type, dynamic>>? __store;
+  Map<Type, StreamSubscription<Action>>? __subscriptions;
   Actions? _actions;
   MonoState(List<StateBase>? reducers) {
     var dispatcher = BehaviorSubject<Action>.seeded(Action(type: '@@INIT'));
     __dispatcher = dispatcher;
-    __store = BehaviorSubject<Map<String, dynamic>>.seeded({});
+    __store = BehaviorSubject<Map<Type, dynamic>>.seeded({});
     _actions = Actions(dispatcher);
-    __subscriptions = <String, StreamSubscription<Action>>{};
+    __subscriptions = <Type, StreamSubscription<Action>>{};
     reducers?.forEach((reducer) {
       registerState(reducer);
     });
@@ -22,72 +22,75 @@ class MonoState {
   BehaviorSubject<Action> get _dispatcher =>
       __dispatcher ?? BehaviorSubject<Action>.seeded(Action(type: '@@INIT'));
 
-  BehaviorSubject<Map<String, dynamic>> get _store =>
-      __store ?? BehaviorSubject<Map<String, dynamic>>.seeded({});
+  BehaviorSubject<Map<Type, dynamic>> get _store =>
+      __store ?? BehaviorSubject<Map<Type, dynamic>>.seeded({});
 
-  Map<String, StreamSubscription<Action>> get _subscriptions =>
-      __subscriptions ?? <String, StreamSubscription<Action>>{};
+  Map<Type, StreamSubscription<Action>> get _subscriptions =>
+      __subscriptions ?? <Type, StreamSubscription<Action>>{};
 
-  Stream<Map<String, dynamic>> get store$ => _store.asBroadcastStream();
+  Stream<Map<Type, dynamic>> get store$ => _store.asBroadcastStream();
   Stream<Action> get dispatcher$ => _dispatcher.asBroadcastStream();
   Actions get action$ => _actions ?? Actions(_dispatcher);
 
   void registerState(StateBase reducer) {
-    if (_store.value.containsKey(reducer.stateName)) {
+    if (_store.value.containsKey(reducer.runtimeType)) {
       return;
     }
     final ns = _newState();
-    ns[reducer.stateName] = reducer.initialState;
+
+    ns[reducer.runtimeType] = reducer.initialState;
     _store.add(ns);
-    dispatch(Action(type: 'registerState(${reducer.stateName})'));
+
     void emitState(dynamic state) {
-      if (_store.value[reducer.stateName] != state) {
+      if (_store.value[reducer.runtimeType] != state) {
         final ns = _newState();
-        ns[reducer.stateName] = state;
+        ns[reducer.runtimeType] = state;
         _store.add(ns);
       }
     }
 
-    _subscriptions[reducer.stateName] = _dispatcher.listen((action) {
+    _subscriptions[reducer.runtimeType] = _dispatcher.listen((action) {
       reducer.mapActionToState(
-          _store.value[reducer.stateName], action, emitState, this);
+          _store.value[reducer.runtimeType], action, emitState, this);
     });
+
+    dispatch(RegisterStateAction('${reducer.runtimeType}'));
   }
 
-  Stream<T> select<T>(String stateName) {
-    return _store.map<T>((dic) => dic[stateName]).distinct();
+  Stream<Model> select<State, Model>() {
+    return _store.map<Model>((dic) => dic[State]).distinct();
   }
 
-  S getState<S>(String stateName) {
-    return _store.value[stateName];
+  dynamic getState<State>() {
+    return _store.value[State];
   }
 
   void dispatch(Action action) {
     _dispatcher.add(action);
   }
 
-  void unregisterState(String stateName) {
-    if (_subscriptions.containsKey(stateName)) {
-      _subscriptions[stateName]?.cancel();
-      _subscriptions.remove(stateName);
+  void unregisterState<State>() {
+    if (_subscriptions.containsKey(State)) {
+      dispatch(UnregisterStateAction('$State'));
+      _subscriptions[State]?.cancel();
+      _subscriptions.remove(State);
       final newState = _newState();
-      newState.remove(stateName);
+      newState.remove(State);
       _store.add(newState);
-      dispatch(Action(type: 'unregisterState($stateName)'));
     }
   }
 
-  void importState(String stateName, dynamic state) {
-    if (_store.value.containsKey(stateName)) {
+  void importState<State>(dynamic state) {
+    if (_store.value.containsKey(State)) {
       final newState = _newState();
-      newState[stateName] = state;
+      newState[State] = state;
       _store.add(newState);
-      dispatch(Action(type: '@importState'));
+      dispatch(ImportStateAction('$State'));
     }
   }
 
-  Map<String, dynamic> _newState() {
-    final map = <String, dynamic>{};
+  Map<Type, dynamic> _newState() {
+    final map = <Type, dynamic>{};
     final oldMap = _store.value;
     for (var key in oldMap.keys) {
       map[key] = oldMap[key];
@@ -95,7 +98,7 @@ class MonoState {
     return map;
   }
 
-  ///It's a clean up function.
+  ///clean up function.
   void dispose() {
     _subscriptions.forEach((key, value) {
       value.cancel();