+++ mono_state/lib/src/stateBase.dart
@@ -4,11 +4,10 @@ import 'monoState.dart';
 typedef EmitCallback<S> = void Function(S state);
 
 abstract class StateBase<S> {
-  final String stateName;
   final S initialState;
 
-  StateBase({required this.stateName, required this.initialState});
+  StateBase(this.initialState);
 
   void mapActionToState(
-      S state, Action action, EmitCallback<S> emit, MonoState store);
+      S state, Action action, EmitCallback<S> emit, MonoState mono);
 }