+++ mono_state/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.0
+
+- Removed stateName props from StateBase class. Now state should be identified by state Type.
+- Function signature changed into - select<State, Model>(), getState<State>()
+
 # 0.2.0
 
 - clear() replaced with dispose()